+++ controllers/database_controller.py
@@ -2,9 +2,9 @@ import sys
 import warlock
 from flask import Flask, request
 from bson.objectid import ObjectId
-from event_service.app import client, db
+from app import client, db
 from .json_controller import *
-from event_service.models.Event import Event
+from models.Event import Event
 
 class MongoDBManager():
